AMP-5901
git-svn-id: file:///svn/trunk/amp@9425 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/helper/ActivityItem.java
@@ -65,7 +65,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
 
 	private ActivityUtil.ActivityAmounts amounts;
     // percent of specific program assigned
-    private Long percent;
+    private Float percent;
     // used only for program percent applying
     private AmpActivity act;
 
@@ -94,7 +94,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
      * @see ActivityItem#ActivityItem(AmpActivity)
 	 */
 
-    public ActivityItem(AmpActivity act,Long percent) {
+    public ActivityItem(AmpActivity act,Float percent) {
         this.act=act;
         this.percent=percent;
 
@@ -110,7 +110,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
 		this(entity,new SimpleDateFormat(Constants.CALENDAR_DATE_FORMAT),"USD",null);
 	}
 
-    public ActivityItem(AmpActivity entity,String curenncyCode,Long percent) throws Exception{
+    public ActivityItem(AmpActivity entity,String curenncyCode,Float percent) throws Exception{
         this(entity,new SimpleDateFormat(Constants.CALENDAR_DATE_FORMAT),curenncyCode,percent);
 	}
 
@@ -120,7 +120,7 @@ public class ActivityItem implements Comparable<ActivityItem>{
 	 * @param entity AmpActivity db entity to construct helper from
 	 * @param frmt date formatter
 	 */
-	public ActivityItem(AmpActivity entity,DateFormat frmt,String curenncyCode,Long percent) throws Exception {
+	public ActivityItem(AmpActivity entity,DateFormat frmt,String curenncyCode,Float percent) throws Exception {
 		if (entity != null) {
 			AmpCategoryValue statusValue = CategoryManagerUtil.getAmpCategoryValueFromListByKey(CategoryConstants.ACTIVITY_STATUS_KEY, entity.getCategories());
 //			statusValue.setValue("fake status");
@@ -309,11 +309,11 @@ public class ActivityItem implements Comparable<ActivityItem>{
 	public void setAmounts(ActivityUtil.ActivityAmounts amounts) {
 		this.amounts = amounts;
 	}
-      public Long getPercent() {
+      public Float getPercent() {
         return percent;
     }
 
-    public void setPercent(Long percent) {
+    public void setPercent(Float percent) {
         this.percent = percent;
     }
